server/requirements.txt
fastapi==0.125.0
uvicorn==0.38.0
python-dotenv==1.2.1
openai==2.13.0
deepgram-sdk==5.3.0
websockets==15.0.1
deepgram-sagemaker==0.2.0
boto3==1.42.13
